==> 1_Balance_Sheet_Analysis.ipynb <==
## Assessing the growth rate of the two companies

In this notebook, we will analyze the growth rate of two companies, Company A and Company B, over a period of 5 years considering some metrics that can be computed from their balance sheets.

# Importing Libraries
import pandas as pd

# Importing the datasets
wmt_balance = pd.read_csv("Dataset\wmt_balance_sheet.csv")
stla_balance = pd.read_csv("Dataset\stla_balance_sheet.csv")
wmt_income = pd.read_csv("Dataset\wmt_income_statement.csv")
stla_income = pd.read_csv("Dataset\stla_income_statement.csv")

# Setting the index to the first column
wmt_balance.set_index(wmt_balance.columns[0], inplace=True)
# Setting the index to the first column
wmt_income.set_index(wmt_income.columns[0], inplace=True)


# Drop last column since it is useless
wmt_balance.drop(wmt_balance.columns[-1], axis=1, inplace=True)


# Setting the index to the first column
stla_balance.set_index(stla_balance.columns[0], inplace=True)

# Drop last column since it is useless
stla_balance.drop(stla_balance.columns[-1], axis=1, inplace=True)

# Setting the index to the first column
stla_income.set_index(stla_income.columns[0], inplace=True)
stla_income = stla_income



# Compute and store important metrics for Walmart
wmt_metrics = pd.DataFrame(columns = wmt_balance.columns)

# Store the number of shares issued in the metrics DataFrame
wmt_metrics.loc["Share Issued"] = wmt_balance.loc["Ordinary Shares Number"]

# Store the total assets in the metrics DataFrame
wmt_metrics.loc["Total Assets"] = wmt_balance.loc["Total Assets"]

# Calculate the average assets using the formula: (Total Assets(t) + Total Assets(t-1)) / 2. If the previous year is not available, use the current year value.
wmt_metrics.loc["Average Assets"] = (wmt_balance.loc["Total Assets"].shift(-1) + wmt_balance.loc["Total Assets"]) / 2
wmt_metrics.loc["Average Assets"].fillna(wmt_balance.loc["Total Assets"], inplace=True)

# Compute the current ratio using the formula: Current Assets / Current Liabilities
wmt_metrics.loc["Current Assets"] = wmt_balance.loc["Current Assets"]
wmt_metrics.loc["Current Liabilities"] = wmt_balance.loc["Current Liabilities"]
wmt_metrics.loc["Total Liabilities"] = wmt_balance.loc["Stockholders Equity"]
wmt_metrics.loc["Total Shareholders' Equity"] = wmt_balance.loc["Total Equity Gross Minority Interest"]
wmt_metrics.loc["Current Ratio"] = wmt_balance.loc["Current Assets"] / wmt_balance.loc["Current Liabilities"]

# Compute the debt to equity ratio using the formula: Total Debt / Stockholders Equity
wmt_metrics.loc["Total Debt"] = wmt_balance.loc["Total Debt"]
wmt_metrics.loc["Stockholders Equity"] = wmt_balance.loc["Stockholders Equity"]
wmt_metrics.loc["Debt to Equity"] = wmt_balance.loc["Total Debt"] / wmt_balance.loc["Total Equity Gross Minority Interest"]

# Store the net income in the metrics DataFrame, as well as the diluted and basic average shares
wmt_metrics.loc["Net Income"] = wmt_income.loc["Net Income From Continuing Operation Net Minority Interest"]
wmt_metrics.loc["Consolidated Net Income"] = wmt_income.loc["Net Income Continuous Operations"]
wmt_metrics.loc["Diluted Average Shares"] = wmt_income.loc["Diluted Average Shares"]
wmt_metrics.loc["Basic Average Shares"] = wmt_income.loc["Basic Average Shares"]

# Compute the Return on Assets (ROA) using the formula: Net Income / Average Assets
wmt_metrics.loc["Return on Assets"] = wmt_metrics.loc["Consolidated Net Income"] / wmt_metrics.loc["Average Assets"]

# Compute Diluted and Basic EPS 
wmt_metrics.loc["Diluted EPS"] = wmt_income.loc["Diluted EPS"]
wmt_metrics.loc["Basic EPS"] = wmt_income.loc["Basic EPS"]

# Store Total Revenue in the metrics DataFrame
wmt_metrics.loc["Total Revenue"] = wmt_income.loc["Total Revenue"]
wmt_metrics.loc["Return on Equity"] = wmt_income.loc["Net Income From Continuing Operation Net Minority Interest"] / wmt_balance.loc["Total Equity Gross Minority Interest"]



# Save the metrics DataFrames to CSV files
wmt_metrics.to_csv("Outputs/wmt_metrics.csv")

# Compute and store important metrics for Stellantis
stla_metrics = pd.DataFrame(columns = stla_balance.columns)

# Store the total assets in the metrics DataFrame
stla_metrics.loc["Total Assets"] = stla_balance.loc["Total Assets"]

# Calculate the market cap using the formula: Share Issued * Price per Share. The prices are imported from Yahoo Finance.
stla_metrics.loc["Market Cap"] = stla_balance.loc["Share Issued"] * 18.18
# Calculate the average assets using the formula: (Total Assets(t) + Total Assets(t-1)) / 2. If the previous year is not available, use the current year value.
stla_metrics.loc["Average Assets"] = (stla_balance.loc["Total Assets"].shift(-1) + stla_balance.loc["Total Assets"]) / 2
stla_metrics.loc["Average Assets"].fillna(stla_balance.loc["Total Assets"], inplace=True)

# Compute the current ratio using the formula: Current Assets / Current Liabilities
stla_metrics.loc["Current Assets"] = stla_balance.loc["Current Assets"]
stla_metrics.loc["Share Issued"] = stla_balance.loc["Share Issued"]
stla_metrics.loc["Current Liabilities"] = stla_balance.loc["Current Liabilities"]
stla_metrics.loc["Total Liabilities"] = stla_balance.loc["Total Liabilities Net Minority Interest"]
stla_metrics.loc["Current Ratio"] = stla_balance.loc["Current Assets"] / stla_balance.loc["Current Liabilities"]

# Compute the debt to equity ratio using the formula: Total Debt / Stockholders Equity
stla_metrics.loc["Total Debt"] = stla_balance.loc["Total Debt"]
stla_metrics.loc["Total Shareholders' Equity"] = stla_balance.loc["Total Equity Gross Minority Interest"]
stla_metrics.loc["Debt to Equity"] = stla_balance.loc["Total Debt"] / stla_balance.loc["Total Equity Gross Minority Interest"]

# Store the net income in the metrics DataFrame, as well as the diluted and basic average shares
stla_metrics.loc["Net Income"] = stla_income.loc["Net Income From Continuing Operation Net Minority Interest"]
stla_metrics.loc["Consolidated Net Income"] = stla_income.loc["Net Income Continuous Operations"]
stla_metrics.loc["Diluted Average Shares"] = stla_income.loc["Diluted Average Shares"]
stla_metrics.loc["Basic Average Shares"] = stla_income.loc["Basic Average Shares"]

# Compute the Return on Assets (ROA) using the formula: Net Income / Average Assets
stla_metrics.loc["Return on Assets"] = stla_metrics.loc["Consolidated Net Income"] / stla_metrics.loc["Average Assets"]

# Compute Diluted and Basic EPS
stla_metrics.loc["Diluted EPS"] = stla_income.loc["Diluted EPS"]
stla_metrics.loc["Basic EPS"] = stla_income.loc["Basic EPS"]

# Store Total Revenue in the metrics DataFrame
stla_metrics.loc["Total Revenue"] = stla_income.loc["Total Revenue"]
stla_metrics.loc["Return on Equity"] = stla_income.loc["Net Income From Continuing Operation Net Minority Interest"] / stla_balance.loc["Total Equity Gross Minority Interest"]

# Save the metrics DataFrames to CSV files
stla_metrics.to_csv("Outputs/stla_metrics.csv")
